Change $scopeSeparator to a space in BitbucketProvider
In Bitbucket's implementation of OAuth2, the space character is the delimiter for scopes.
When trying to request multiple scopes, the request fails with the callback url having `error_description` appended to it with the value of 'Unknown scope: ', and then a comma-delimited list of scopes.  After that the listed `error` variable has a value of `invalid_scope`.